+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Added `LVMActor` class with actor state and troubleshooting framework.
 
+
 ## 0.1.0 - January 19, 2024
 
 ### 🚀 New